Moved apiManagementResourceHostnameSchema and apiManagementResourceHostnameProxySchema to separate file
