+++ azurerm/internal/services/apimanagement/api_management_resource.go
@@ -1180,54 +1180,6 @@ func flattenApiManagementVirtualNetworkConfiguration(input *apimanagement.Virtua
 	return []interface{}{virtualNetworkConfiguration}
 }
 
-func apiManagementResourceHostnameSchema() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"host_name": {
-			Type:         schema.TypeString,
-			Required:     true,
-			ValidateFunc: validation.StringIsNotEmpty,
-		},
-
-		"key_vault_id": {
-			Type:         schema.TypeString,
-			Optional:     true,
-			ValidateFunc: azure.ValidateKeyVaultChildIdVersionOptional,
-		},
-
-		"certificate": {
-			Type:         schema.TypeString,
-			Optional:     true,
-			Sensitive:    true,
-			ValidateFunc: validation.StringIsNotEmpty,
-		},
-
-		"certificate_password": {
-			Type:         schema.TypeString,
-			Optional:     true,
-			Sensitive:    true,
-			ValidateFunc: validation.StringIsNotEmpty,
-		},
-
-		"negotiate_client_certificate": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Default:  false,
-		},
-	}
-}
-
-func apiManagementResourceHostnameProxySchema() map[string]*schema.Schema {
-	hostnameSchema := apiManagementResourceHostnameSchema()
-
-	hostnameSchema["default_ssl_binding"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-		Computed: true, // Azure has certain logic to set this, which we cannot predict
-	}
-
-	return hostnameSchema
-}
-
 func parseApiManagementNilableDictionary(input map[string]*string, key string) bool {
 	log.Printf("Parsing value for %q", key)
